extract authentication test to separate test project
