+++ OpenRiaServices.DomainServices.Server.Authentication.AspNetMembership/Framework/AuthenticationBase.cs
