+++ OpenRiaServices.DomainServices.Server.Authentication.AspNetMembership/Framework/UserBase.cs
