+++ OpenRiaServices.DomainServices.Server.Authentication.AspNetMembership/Test/App.config
@@ -1,0 +1,37 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <system.web>
+    <membership defaultProvider="MockMembershipProvider">
+      <providers>
+        <add name="MockMembershipProvider" type="OpenRiaServices.DomainServices.Server.ApplicationServices.Test.MockMembershipProvider, OpenRiaServices.DomainServices.Server.Authentication.AspNetMembership.Test" />
+      </providers>
+    </membership>
+    <roleManager defaultProvider="MockRoleProvider" enabled="true">
+      <providers>
+        <add name="MockRoleProvider" type="OpenRiaServices.DomainServices.Server.ApplicationServices.Test.MockRoleProvider, OpenRiaServices.DomainServices.Server.Authentication.AspNetMembership.Test" />
+      </providers>
+    </roleManager>
+    <profile defaultProvider="MockProfileProvider" enabled="true" automaticSaveEnabled="false">
+      <providers>
+        <add name="MockProfileProvider" type="OpenRiaServices.DomainServices.Server.ApplicationServices.Test.MockProfileProvider, OpenRiaServices.DomainServices.Server.Authentication.AspNetMembership.Test" />
+      </providers>
+      <properties>
+        <add name="UserBoolean" type="bool" />
+        <add name="UserDouble" type="double" />
+        <add name="UserInt32" type="int" />
+        <add name="UserString" type="string" />
+        <add name="AliasedString" type="string" />
+        <add name="UserStringReadOnly" type="string" readOnly="true" />
+        <add name="UserStringNotReadOnly" type="string" />
+        <add name="VirtualNotAliased" type="string" />
+        <add name="VirtualInProfile" type="string" />
+        <add name="VirtualReadOnly" type="string" />
+        <add name="VirtualNotReadOnly" type="string" />
+      </properties>
+    </profile>
+    <authentication mode="Forms" />
+  </system.web>
+  <startup>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+  </startup>
+</configuration>
