+++ OpenRiaServices.DomainServices.Server.Authentication.AspNetMembership/Test/MockUser.cs
