+++ OpenRiaServices.DomainServices.Server.Authentication.AspNetMembership/Test/UserBaseTest.cs
