+++ OpenRiaServices.DomainServices.Server/Test/App.config
@@ -1,8 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <configSections>
-    <section name="msbuildToolsets" type="Microsoft.Build.BuildEngine.ToolsetConfigurationSection, Microsoft.Build.Engine, Version=3.5.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
-    
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
   <!-- For more information on Entity Framework configuration, visit http://go.microsoft.com/fwlink/?LinkID=237468 --></configSections>
   <connectionStrings>
@@ -10,50 +8,6 @@
     <add name="NorthwindEntities" connectionString="metadata=res://*/DataModels.Northwind.csdl|res://*/DataModels.Northwind.ssdl|res://*/DataModels.Northwind.msl;provider=System.Data.SqlClient;provider connection string=&quot;Data Source=(localdb)\MSSQLLocalDB;Initial Catalog=Northwind;Integrated Security=True;MultipleActiveResultSets=True&quot;" providerName="System.Data.EntityClient" />
     <add name="NorthwindPOCOEntities" connectionString="metadata=res://*/Northwind.csdl|res://*/Northwind.ssdl|res://*/Northwind.msl;provider=System.Data.SqlClient;provider connection string=&quot;Data Source=(localdb)\MSSQLLocalDB;Initial Catalog=Northwind;Integrated Security=True;MultipleActiveResultSets=True&quot;" providerName="System.Data.EntityClient" />
   </connectionStrings>
-  <runtime>
-    <generatePublisherEvidence enabled="false" />
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="Microsoft.Build.Framework" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-99.9.9.9" newVersion="4.0.0.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Microsoft.CompactFramework.Build.Tasks" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-99.9.9.9" newVersion="10.0.0.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
-  <system.web>
-    <membership defaultProvider="MockMembershipProvider">
-      <providers>
-        <add name="MockMembershipProvider" type="OpenRiaServices.DomainServices.Server.ApplicationServices.Test.MockMembershipProvider, OpenRiaServices.DomainServices.Server.Test" />
-      </providers>
-    </membership>
-    <roleManager defaultProvider="MockRoleProvider" enabled="true">
-      <providers>
-        <add name="MockRoleProvider" type="OpenRiaServices.DomainServices.Server.ApplicationServices.Test.MockRoleProvider, OpenRiaServices.DomainServices.Server.Test" />
-      </providers>
-    </roleManager>
-    <profile defaultProvider="MockProfileProvider" enabled="true" automaticSaveEnabled="false">
-      <providers>
-        <add name="MockProfileProvider" type="OpenRiaServices.DomainServices.Server.ApplicationServices.Test.MockProfileProvider, OpenRiaServices.DomainServices.Server.Test" />
-      </providers>
-      <properties>
-        <add name="UserBoolean" type="bool" />
-        <add name="UserDouble" type="double" />
-        <add name="UserInt32" type="int" />
-        <add name="UserString" type="string" />
-        <add name="AliasedString" type="string" />
-        <add name="UserStringReadOnly" type="string" readOnly="true" />
-        <add name="UserStringNotReadOnly" type="string" />
-        <add name="VirtualNotAliased" type="string" />
-        <add name="VirtualInProfile" type="string" />
-        <add name="VirtualReadOnly" type="string" />
-        <add name="VirtualNotReadOnly" type="string" />
-      </properties>
-    </profile>
-    <authentication mode="Forms" />
-  </system.web>
   <entityFramework>
     <defaultConnectionFactory type="System.Data.Entity.Infrastructure.LocalDbConnectionFactory, EntityFramework">
       <parameters>
