+++ OpenRiaServices.DomainServices.Server/Test/DomainServiceTests.cs
@@ -7,7 +7,6 @@ using System.Data.Linq;
 using System.Globalization;
 using System.Linq;
 using System.Reflection;
-using System.Security.Principal;
 using OpenRiaServices.DomainServices.Client.Test;
 using OpenRiaServices.DomainServices.EntityFramework;
 using OpenRiaServices.DomainServices.Hosting;
@@ -2085,96 +2084,6 @@ namespace OpenRiaServices.DomainServices.Server.Test
         {
             return base.ExecuteChangeSet();
         }
-    }
-
-    public class MockDataService : IServiceProvider
-    {
-        private readonly IPrincipal user;
-
-        public MockDataService(IPrincipal user)
-        {
-            this.user = user;
-        }
-
-        #region IServiceProvider Members
-
-        public object GetService(Type serviceType)
-        {
-            if (serviceType == typeof(IPrincipal))
-            {
-                return this.user;
-            }
-            return null;
-        }
-
-        #endregion
-    }
-
-    public class MockUser : IPrincipal, IIdentity
-    {
-        private readonly IEnumerable<string> roles;
-        private readonly string name;
-        private bool isAuthenticated;
-
-        public MockUser(string name)
-        {
-            this.name = name;
-        }
-
-        public MockUser(string name, IEnumerable<string> roles)
-        {
-            this.name = name;
-            this.roles = roles;
-        }
-
-        #region IPrincipal Members
-
-        public IIdentity Identity
-        {
-            get
-            {
-                return this;
-            }
-        }
-
-        public bool IsInRole(string role)
-        {
-            return this.roles.Contains(role);
-        }
-
-        #endregion
-
-        #region IIdentity Members
-
-        public string AuthenticationType
-        {
-            get
-            {
-                return "forms";
-            }
-        }
-
-        public bool IsAuthenticated
-        {
-            get
-            {
-                return this.isAuthenticated;
-            }
-            set
-            {
-                this.isAuthenticated = value;
-            }
-        }
-
-        public string Name
-        {
-            get
-            {
-                return this.name;
-            }
-        }
-
-        #endregion
     }
 
     public class ReadonlyEntity
